pyproject.toml
[tool.poetry]
name = "anavara task"
version = "0.1.0"
description = "Interview Challenge"
authors = ["Oluwaseyi Daniel <dexter_daniel@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
django = "^4.2"
pydantic = {version = "^2.3", extras = ["email"]}
pydantic-settings = "^2.0.3"
djangorestframework-simplejwt = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
